+++ functions/sdk/index.d.ts
@@ -45,6 +45,9 @@ declare module "@teamkeel/sdk/index" {
 
   //@ts-ignore
   export * from "@teamkeel/client";
+
+  export * from "@teamkeel/sdk/returnTypes";
+
   export {
     Query,
     QueryConstraints,
@@ -97,15 +100,16 @@ declare module "@teamkeel/sdk/query" {
     Input,
     OrderClauses,
   } from "@teamkeel/sdk/types";
+  import * as ReturnTypes from "@teamkeel/sdk/returnTypes";
   export class ChainableQuery<T> {
     private readonly tableName;
     private readonly conditions;
     private readonly pool;
     constructor({ tableName, pool, conditions }: ChainedQueryOpts<T>);
     orWhere: (conditions: Conditions<T>) => ChainableQuery<T>;
-    all: () => Promise<T[]>;
+    all: () => Promise<ReturnTypes.FunctionListResponse<T>>;
     order: (clauses: OrderClauses<T>) => ChainableQuery<T>;
-    findOne: () => Promise<T>;
+    findOne: () => Promise<ReturnTypes.FunctionGetResponse<T>>;
     sql: ({ asAst }: SqlOptions) => string | TaggedTemplateLiteralInvocation<T>;
     private appendConditions;
   }
@@ -115,12 +119,19 @@ declare module "@teamkeel/sdk/query" {
     private orderClauses;
     private readonly pool;
     constructor({ tableName, pool, logger }: QueryOpts);
-    create: (inputs: Partial<T>) => Promise<T>;
+    create: (
+      inputs: Partial<T>
+    ) => Promise<ReturnTypes.FunctionCreateResponse<T>>;
     where: (conditions: Conditions<T>) => ChainableQuery<T>;
-    delete: (id: string) => Promise<boolean>;
-    findOne: (conditions: Conditions<T>) => Promise<T>;
-    update: (id: string, inputs: Input<T>) => Promise<T>;
-    all: () => Promise<T[]>;
+    delete: (id: string) => Promise<ReturnTypes.FunctionDeleteResponse<T>>;
+    findOne: (
+      conditions: Conditions<T>
+    ) => Promise<ReturnTypes.FunctionGetResponse<T>>;
+    update: (
+      id: string,
+      inputs: Input<T>
+    ) => Promise<ReturnTypes.FunctionUpdateResponse<T>>;
+    all: () => Promise<ReturnTypes.FunctionListResponse<T>>;
   }
   export {};
 }
@@ -182,6 +193,53 @@ declare module "@teamkeel/sdk/types" {
     email: string;
   }
 }
+
+declare module "@teamkeel/sdk/returnTypes" {
+  // ValidationErrors will be returned when interacting with
+  // the Query API (creating, updating entities)
+  export interface ValidationError {
+    field: string;
+    message: string;
+    code: string;
+  }
+
+  // ExecutionError represents other misc errors
+  // that can occur during the execution of a custom function
+  export interface ExecutionError {
+    message: string;
+
+    // todo: implement stacks
+    stack: string;
+  }
+
+  export type FunctionError = ValidationError | ExecutionError;
+
+  export interface FunctionCreateResponse<T> {
+    object?: T;
+    errors?: FunctionError[];
+  }
+
+  export interface FunctionGetResponse<T> {
+    object?: T;
+    errors?: FunctionError[];
+  }
+
+  export interface FunctionDeleteResponse<T> {
+    success: boolean;
+  }
+
+  export interface FunctionListResponse<T> {
+    collection: T[];
+
+    // todo: add type for pagination
+  }
+
+  export interface FunctionUpdateResponse<T> {
+    object?: T;
+    errors?: FunctionError[];
+  }
+}
+
 declare module "@teamkeel/sdk" {
   import main = require("@teamkeel/sdk/index");
   export = main;