+++ integration/testdata/built_in_actions/example.test.ts
@@ -1,13 +1,13 @@
 import { test, expect, Actions } from '@teamkeel/testing'
 
 test('create action', async () => {
-  const { result } = await Actions.createPost({ title: 'foo' })
-  expect.equal(result.title, 'foo')
+  const { object: createdPost } = await Actions.createPost({ title: 'foo' })
+  expect.equal(createdPost.title, 'foo')
 })
 
 test('get action', async () => {
-  const { result: post } = await Actions.createPost({ title: 'foo' })
-  const { result } = await Actions.getPost({ id: post.id })
+  const { object: post } = await Actions.createPost({ title: 'foo' })
+  const { object: fetchedPost } = await Actions.getPost({ id: post.id })
 
-  expect.equal(result.id, post.id)
+  expect.equal(fetchedPost.id, post.id)
 })