+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [0.105.0](https://github.com/teamkeel/keel/compare/v0.104.1...v0.105.0) (2022-08-30)
+
+
+### Features
+
+* built-in authenticate() operation with graphql support ([#328](https://github.com/teamkeel/keel/issues/328)) ([f7b76fa](https://github.com/teamkeel/keel/commit/f7b76fa3219652d804f6105d85347bcabbaed103))
+
+## [0.104.1](https://github.com/teamkeel/keel/compare/v0.104.0...v0.104.1) (2022-08-30)
+
+
+### Bug Fixes
+
+* update sdk type defs to reflect new query api return types ([#330](https://github.com/teamkeel/keel/issues/330)) ([67dedb9](https://github.com/teamkeel/keel/commit/67dedb9263f7b0b3601df1c5ed7aff64b68ab719))
+
+# [0.104.0](https://github.com/teamkeel/keel/compare/v0.103.0...v0.104.0) (2022-08-30)
+
+
+### Features
+
+* update codegenned wrapper func to use new return types from sdk ([#329](https://github.com/teamkeel/keel/issues/329)) ([8467c93](https://github.com/teamkeel/keel/commit/8467c9394b988abfccb8fdbeb172ae237a0dbbfd))
+
+# [0.103.0](https://github.com/teamkeel/keel/compare/v0.102.21...v0.103.0) (2022-08-26)
+
+
+### Features
+
+* add custom function return type interfaces ([#320](https://github.com/teamkeel/keel/issues/320)) ([8aa8ddb](https://github.com/teamkeel/keel/commit/8aa8ddb529fc6021d0e479e786a2473adbcf90e2))
+
 ## [0.102.21](https://github.com/teamkeel/keel/compare/v0.102.20...v0.102.21) (2022-08-26)